Created HashPartitionedReadWriteStorage
Also revised benchmarking to include multiple time points

+++ src/main/java/dbf0/disk_key_value/readwrite/btree/LockingBlockBTree.java
@@ -53,7 +53,7 @@ public class LockingBlockBTree<K extends Comparable<K>, V> implements BTree<K, V
     tree.initialize();
   }
 
-  @Override public int size() throws IOException {
+  @Override public long size() throws IOException {
     return lock.callWithReadLock(tree::size);
   }
